fix(test_hierarchical_heatmap): fix cell count test failure
Updates the unit test for hierarchical_heatmap. show_counts argument in the hierarchical_heatmap function is set to True so that the y-axis labels including cell counts are enabled. Previously, the test relied on the default show_counts=False, resulting in labels without cell counts which caused the test to fail.